+++ src/cli.rs
@@ -13,11 +13,9 @@
 // limitations under the License.
 
 use std::path::{Path, PathBuf};
-use std::{env::current_dir, error::Error};
 
 use clap::Parser;
 use directories::UserDirs;
-use path_clean::PathClean;
 
 use crate::{sampling::SamplingFilter, ImageFormat};
 
@@ -49,59 +47,30 @@ pub struct Args {
     pub sampling_filter: SamplingFilter,
 }
 
-/// Represents CLI arguments parsed by `clap` with reasonable
-/// defaults for arguments whose parsed values are *None*.
-#[derive(Debug)]
-pub struct NormalizedArgs {
-    /// Local image file path for generating the thumbnail
-    pub path: Box<Path>,
-
-    /// Generated thumbnail name.
-    pub out_name: String,
-
-    /// Specifies the output directory for the thumbnail.
-    pub out_dir: PathBuf,
-
-    /// The thumbnail's output format.
-    pub format: ImageFormat,
-
-    /// Sampling algorithm to use for thumbnail generation.
-    pub sampling_filter: SamplingFilter,
-}
-
 impl Args {
-    /// Returns parsed arguments with reasonable defaults.
-    pub fn normalize(self) -> NormalizedArgs {
-        let out_name = self.out_name.unwrap_or_else(|| {
-            let p_without_ext = self.path.with_extension("");
-            let original_filename = p_without_ext.file_name().unwrap();
-
-            if original_filename.is_empty() {
-                panic!("Invalid empty path to image file")
-            }
-
-            format!(
-                "{}_thumb.{}",
-                &original_filename.to_str().unwrap(),
-                &self.format
-            )
-        });
-
-        let path = abs_path_to_img(&self.path).unwrap();
-
-        NormalizedArgs {
-            path,
-            out_name,
-            out_dir: self.out_dir,
-            format: self.format,
-            sampling_filter: self.sampling_filter,
-        }
-    }
-}
-
-impl NormalizedArgs {
     pub fn get_final_output_path(&self) -> PathBuf {
-        self.out_dir.join(&self.out_name)
+        self.out_dir.join(self.get_out_name())
+    }
+
+    fn get_out_name(&self) -> String {
+        self.out_name
+            .as_ref()
+            .map(|val| val.to_string())
+            .unwrap_or_else(|| {
+                let p_without_ext = self.path.with_extension("");
+                let original_filename = p_without_ext.file_name().unwrap();
+
+                if original_filename.is_empty() {
+                    panic!("Invalid empty path to image file")
+                }
+
+                format!(
+                    "{}_thumb.{}",
+                    &original_filename.to_str().unwrap(),
+                    &self.format
+                )
+                .to_string()
+            })
     }
 }
 
@@ -115,17 +84,4 @@ fn default_output_dir() -> PathBuf {
             panic!("Unable to find user 'Pictures' directory in the operating sytem")
         })
         .to_path_buf()
-}
-
-fn abs_path_to_img(path: &impl AsRef<Path>) -> Result<Box<Path>, Box<dyn Error>> {
-    let path = path.as_ref();
-
-    let absolute_path = if path.is_absolute() {
-        path.to_path_buf().into_boxed_path()
-    } else {
-        current_dir()?.join(path).into_boxed_path()
-    }
-    .clean();
-
-    Ok(absolute_path.into_boxed_path())
 }
